Add files via upload
This project features user registration and login forms built with Visual Studio using Windows Forms for the interface. It includes buttons for users to log in if already registered or redirect to a registration form if they don’t have an account. The user-friendly navigation provides a smooth onboarding process, making account management straightforward.