Add files via upload
This project includes a user login form created with Visual Studio using Windows Forms for the interface. Designed to facilitate secure and easy access, the form enables registered users to log in, with options to navigate to registration if an account has not been set up. The interface is built for intuitive and smooth user authentication.